src/main/java/org/bitbiome/commands/MapCommand.java
package org.bitbiome.commands;
import org.bitbiome.classes.TravelEngine;
import org.json.JSONArray;
import java.util.Scanner;

public class MapCommand implements CommandAPI{
    @Override
    public void performCommand(Scanner scanner, boolean isRunning, String message, TravelEngine travelEngine) {
        System.out.println(getMapMessage(travelEngine));
    }
    public static String getMapMessage(TravelEngine travelEngine) {
        StringBuilder outputMessage = new StringBuilder();
        JSONArray locations = travelEngine.getLocationList();
            outputMessage
                    .append("Map:\n")
                    .append("Wueste         Gruenland       Winterland\n\n")
                    .append("Strand         Wald            Berge\n\n");
        return outputMessage.toString();
    }

}
